requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cloudinary==1.39.0
slowapi==0.1.9
pymysql==1.1.0
email-validator==2.1.1